Apply discount code for main group level discounts
This change, together with the change in the PMPro plugin uses the jQuery triggered event to update the seat cost (applies the discount code over the seat cost as well) after a discount code is applied on the main group level. 
It also calculates the discount on seats based on the request (if it is a response from a submission or if it is an URL that already comes with the discount code applied).